+++ angular-workspace/projects/ni/nimble-angular/src/directives/text-field/nimble-text-field.directive.ts
@@ -1,7 +1,7 @@
 import { Directive, ElementRef, Input, Renderer2 } from '@angular/core';
 import type { TextField } from '@ni/nimble-components/dist/esm/text-field';
 import { TextFieldAppearance, TextFieldType } from '@ni/nimble-components/dist/esm/text-field/types';
-import { BooleanValueOrAttribute, toBooleanProperty } from '../utilities/template-value-helpers';
+import { BooleanValueOrAttribute, toBooleanProperty } from '@ni/nimble-angular/internal-utilities';
 
 export type { TextField };
 export { TextFieldType, TextFieldAppearance };